% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-binommix-sample.R
\docType{data}
\name{binommix_sample}
\alias{binommix_sample}
\title{A simulated Mixture Binomial dataset}
\format{
A data frame with 100 rows and 2 variables:
\describe{
  \item{x}{the number of successes in each set of Bernoulli trails}
  \item{n}{total Bernoulli trails corresponding to x}
}
}
\usage{
binommix_sample
}
\description{
A simulated Mixture Binomial dataset based on rBinomMix algorithm
}
\keyword{datasets}
