% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binommixEM.R
\name{binommixEM}
\alias{binommixEM}
\title{EM algorithm to estimate mixture binomial}
\usage{
binommixEM(
  x,
  n,
  k,
  pi_inits = rep(1/k, k),
  p_inits = runif(k, 0, 1),
  maxit = 1000,
  tol = .Machine$double.eps
)
}
\arguments{
\item{x}{a numeric vector of observed successes in each set of Bernoulli trails}

\item{n}{a numeric vector of total Bernoulli trails corresponding to x}

\item{k}{a numeric value specific the number of binomial mixtures or clusters}

\item{pi_inits}{a numeric vector of initial values in probability assigned to each binomial distribution}

\item{p_inits}{a numeric vector of initial values in probability of success in each of the binomial distribution}

\item{maxit}{max iterations}

\item{tol}{stopping criteria}
}
\value{
\itemize{
  \item{pi} {The assignment probability for mixture binomial distributions.}
  \item{p} {The probability parameters for mixture binomial distributions.}
  \item{cluster_ind} {The cluster assignment matrix, 0 or 1.}
  \item{iter} {The number of iterations performed.}
  \item{loglikelihood} {The vector of log-likelihood in each iteration.}
  \item{x} {The mixture binomial observed successes.}
  \item{size} {The mixture binomial Bernoulli tails.}
  \item{soft_label} {The vector of numeric matrix of final assignment probability for each observation.}
}
}
\description{
EM algorithm to estimate mixture binomial
}
\examples{
u_x <- rBinomMix(1000, alpha=c(0.4, 0.6), theta=c(0.3, 0.8), size=30); u_n <- rep(30, 1000)
binommixEM(x=u_x, n=u_n, k=2, pi_inits=c(0.3, 0.7), p_inits=c(0.4, 0.2))
}
