% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binommixMstep.R
\name{binommixMstep}
\alias{binommixMstep}
\title{M step in estimating mixture binomial}
\usage{
binommixMstep(x, n, e.step)
}
\arguments{
\item{x}{a numeric vector of observed successes in each set of Bernoulli trails}

\item{n}{a numeric vector of total Bernoulli trails corresponding to x}

\item{e.step}{output from binommixEstep, a numeric matrix of assignment probability for observed successes in each set of Bernoulli trails}
}
\value{
\itemize{
  \item{pi} {The expected assignment probability for mixture binomial distributions.}
  \item{p} {The probability parameters for mixture binomial distributions.}
}
}
\description{
M step in estimating mixture binomial
}
\examples{
u_x <- rBinomMix(1000, alpha=c(0.4, 0.6), theta=c(0.3, 0.8), size=30); u_n <- rep(30, 1000)
estep <- binommixEstep(x=u_x, n=u_n, pi=c(0.5, 0.5), p=c(0.2, 0.4))
binommixMstep(x=u_x, n=u_n, e.step=estep)
}
