% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binommixPlot.R
\name{binommixPlot}
\alias{binommixPlot}
\title{Visualization of mixture binomial result}
\usage{
binommixPlot(binommixem)
}
\arguments{
\item{binommixem}{an output from binommixEM}
}
\value{
a plot visualizing mixture binomial result
}
\description{
Visualization of mixture binomial result
}
\examples{
u_x <- rBinomMix(1000, alpha=c(0.4, 0.6), theta=c(0.3, 0.8), size=30) ;u_n <- rep(30, 1000)
test_case <- binommixEM(u_x, u_n, k=2, pi_inits=c(0.3, 0.7), p_inits=c(0.4, 0.2))
fit_plot <- binommixPlot(test_case)
}
