% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnormal_EM.R
\name{mixnormal_EM}
\alias{mixnormal_EM}
\title{EM algorithm of mixture normal distribution}
\usage{
mixnormal_EM(
  y,
  theta_vector = NULL,
  mu_vector = NULL,
  sigma_vector = NULL,
  small = 1e-05,
  max_iter = 1000
)
}
\arguments{
\item{y}{A numeric vector of mixture normal observations.}

\item{theta_vector}{A numeric vector indicating proportions of mixing components.}

\item{mu_vector}{A numeric vector indicating mean of mixing components.}

\item{sigma_vector}{A numeric vector indicating standard deviation of mixing components.}

\item{small}{A number indicating the criterion for defining convergence. The default value is 1e-8.}

\item{max_iter}{A number indicating max iteration count. The default value is 1000.}
}
\value{
\itemize{
  \item y - The mixture normal observations.
  \item theta - The vector of proportions of mixing components at the last iteration.
  \item mu - The vector of mean of mixing components at the last iteration.
  \item sigma - The vector of standard deviation of mixing components at the last iteration.
  \item iteration.count - The number of iterations performed.
  \item converge - A boolean indicator of whether the algorithm reach converged.
}
}
\description{
EM algorithm of mixture normal distribution
}
\examples{
y <- c(8.1, 8.2, 8.1, 8.2, 8.2, 7.4, 7.3, 7.4, 8.1, 8.1, 7.9, 7.8, 8.2, 7.9, 7.9, 8.1, 8.1)
mixnormal_EM(y,theta_vector = c(0.2, 0.8),mu_vector = c(0.5, 2), sigma_vector = c(2, 1))
}
