% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-normalmix-sample.R
\docType{data}
\name{mixnormal_sample}
\alias{mixnormal_sample}
\title{A simulated Mixture Normal dataset}
\format{
A data frame with 2 dataframes:
\describe{
  \item{trueParams}{The weignt, mean and standard deviation of mixture components}
  \item{simObs}{500 mixture normal sample}
}
}
\usage{
mixnormal_sample
}
\description{
A simulated Mixture Normal dataset based on simu_mixnormal function
}
\keyword{datasets}
