% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix_plot.R
\name{normalmix_plot}
\alias{normalmix_plot}
\title{Visualization of EM optimization result of mixture of normal}
\usage{
normalmix_plot(normalmixem)
}
\arguments{
\item{normalmixem}{The object of EM optimization result (The return object of function
mixnormal_EM).}
}
\value{
A plot of mixture normal distributions
}
\description{
Visualization of EM optimization result of mixture of normal
}
\examples{
simu_x <- simu_mixnormal(1000)
test_case <- mixnormal_EM(simu_x$simObs$normObs,theta_vector = c(0.2, 0.8),
mu_vector = c(0.5, 2),sigma_vector = c(2, 1))
normalmix_plot(test_case)
}
