% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pois_lambda_init.R
\docType{data}
\name{pois_lambda_init}
\alias{pois_lambda_init}
\title{initial lambda vector.}
\format{
A vector with 3 initial lambdas
\describe{
  \item{lambda}{3 lambdas}
}
}
\usage{
pois_lambda_init
}
\description{
A data set containing the simulated initial lambda vector.
}
\keyword{datasets}
