% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pois_pi_init.R
\docType{data}
\name{pois_pi_init}
\alias{pois_pi_init}
\title{initial Pi vector.}
\format{
A vector with 3 initial pi
\describe{
  \item{pi}{3 pi}
}
}
\usage{
pois_pi_init
}
\description{
A data set containing the simulated initial lambda vector.
}
\keyword{datasets}
