% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonMixPlot.R
\name{poismixPlot}
\alias{poismixPlot}
\title{plotting original observation mixture Poisson data with after clustering}
\usage{
poismixPlot(poismixem)
}
\arguments{
\item{poismixem}{output from the function poissonEM in mixtureEM package}
}
\value{
a plot visualizing mixture poisson result
}
\description{
plotting original observation mixture Poisson data with after clustering
}
