% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonEstep.R
\name{poissonEstep}
\alias{poissonEstep}
\title{Calculating the Q function and log-likelihood of Mixture Poisson for every step}
\usage{
poissonEstep(data, pi, lambda)
}
\arguments{
\item{data}{a vector that stored the observed value data.}

\item{pi}{a vector that stored the probability of each component Poisson model from the last iteration.}

\item{lambda}{a vector that stored the mean parameter of each component Poisson model from the last iteration.}
}
\value{
\itemize{
\item{z} The 'z' value matrix which will be used in M-step.
\item{llike} The vector of log likelihood value
}
}
\description{
Calculating the Q function and log-likelihood of Mixture Poisson for every step
}
\examples{
x = c(94, 89, 94, 28, 25, 24, 34, 33, 34, 24, 33, 29, 35, 30, 29, 27, 24, 28)
poissonEstep(data = x, pi = c(0.5, 0.5), lambda = c(80,30))
}
