% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonMixEM.R
\name{poissonMixEM}
\alias{poissonMixEM}
\title{Finding MLE of mixture Poisson models}
\usage{
poissonMixEM(data, pi, lambda, small = 1e-04, maxIter = 1000)
}
\arguments{
\item{data}{a vector that stored observed data}

\item{pi}{a vector that stored the probability of each component from the last iteration}

\item{lambda}{a vector that stored the mean parameter of each component from the last iteration}

\item{small}{a numerical value that controlling the stopping criteria}

\item{maxIter}{a numerical value that limiting the maximum iterations}
}
\value{
\itemize{
\item{pi} {The assignment probability for mixture poisson distributions.}
\item{lambda} {The lambda parameter in the component poisson distributions.}
\item{x} {The mixture poisson observed successes.}
\item{z} {The 'z' value matrix which will be used in M-step.}
\item{iteration} {The number of iterations performed.}
\item{log-likelihood} {The vector of log likelihood value.}
}
}
\description{
Finding MLE of mixture Poisson models
}
\examples{
x = c(94, 89, 94, 28, 25, 24, 34, 33, 34, 24, 33, 29, 35, 30, 29, 27, 24, 28)
pi.init = pois_pi_init
lambda.init = pois_lambda_init
poissonMixEM(x, pi.init,lambda.init)
}
