% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonMstepLambda.R
\name{poissonMlambda}
\alias{poissonMlambda}
\title{Optimize M step with parameter Lambda}
\usage{
poissonMlambda(z, data)
}
\arguments{
\item{z}{a matrix got from E step}

\item{data}{a vector that stored the observed value data}
}
\value{
lambda The vector of lambda for the components in the mixture poisson distribution.
}
\description{
Optimize M step with parameter Lambda
}
\examples{
X = c(94, 89, 94, 28, 25, 24, 34, 33, 34, 24, 33, 29, 35, 30, 29, 27, 24, 28)
Estep = poissonEstep(data = X, pi = c(0.5, 0.5), lambda = c(80,30))$z
poissonMlambda(Estep , data = X)
}
