% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonMstepPi.R
\name{poissonMpi}
\alias{poissonMpi}
\title{Optimize M step with parameter Pi}
\usage{
poissonMpi(z)
}
\arguments{
\item{z}{a matrix got from E step}
}
\value{
pi {The assignment probability for mixture poisson distributions.}
}
\description{
Optimize M step with parameter Pi
}
\examples{
x = c(94, 89, 94, 28, 25, 24, 34, 33, 34, 24, 33, 29, 35, 30, 29, 27, 24, 28)
Estep = poissonEstep(data = x, pi = c(0.5, 0.5), lambda = c(80,30))$z
poissonMpi(Estep)
}
