% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBinomMix.R
\name{rBinomMix}
\alias{rBinomMix}
\title{Simulate mixture Binomial data}
\usage{
rBinomMix(n, alpha, theta, size)
}
\arguments{
\item{n}{a numeric value specify the total sample size}

\item{alpha}{a numeric vector to specify assigned probabilities for each of the binomial distribution, should sum to 1}

\item{theta}{a numeric vector of p parameters specified for each of the binomial distributions}

\item{size}{a numeric value specify the total Bernoulli trails}
}
\value{
a vector of n simulated successes following a mixture binomial distribution
}
\description{
Simulate mixture Binomial data
}
\examples{
rBinomMix(1000, alpha=c(0.4, 0.6), theta=c(0.3, 0.8), size=30)
}
