% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPoisson.R
\name{simPoisson}
\alias{simPoisson}
\title{Title simulating a mixture poisson model}
\usage{
simPoisson(n = 1000, k = 4, sep = 0.01)
}
\arguments{
\item{n}{number of observations}

\item{k}{number of components}

\item{sep}{amount of overlap between components (small sep = well-separated clusters)}
}
\value{
dataP a vector of simulated mixture poisson distribution
}
\description{
Title simulating a mixture poisson model
}
\examples{
simPoisson(35, 3)

}
