% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_mixnormal.R
\name{simu_mixnormal}
\alias{simu_mixnormal}
\title{Simulate Mixtures of Normal Distribution}
\usage{
simu_mixnormal(
  n,
  ncluster = NULL,
  theta_vector = NULL,
  mu_vector = NULL,
  sigma_vector = NULL
)
}
\arguments{
\item{n}{A number indicating number of observations you want to simulate}

\item{ncluster}{A number indicating number of mixing components. The default value is 2}

\item{theta_vector}{A numeric vector indicating proportions of mixing components. The default value is (0.3,0.7)}

\item{mu_vector}{A numeric vector indicating mean of mixing components. The default value is (1,3)}

\item{sigma_vector}{A numeric vector indicating standard deviation of mixing components. The default value is (3,0.4)}
}
\value{
A data frame with parameters and simulated observations
}
\description{
Simulate Mixtures of Normal Distribution
}
\examples{
simu_mixnormal(100)
}
