% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_test_pois.R
\docType{data}
\name{test_pois}
\alias{test_pois}
\title{simulated Poisson mixture model dataset}
\format{
A vector with 3 initial pi
\describe{
  \item{data}{1000 random data points from 3 mixture Poisson models}
}
}
\usage{
test_pois
}
\description{
A data set containing 1000 randomly simulated mixture Poisson models
}
\keyword{datasets}
